Change pink to green
"squishy pink newbie" was never intended to imply any variety of human but it was pointed out that this might be unwelcoming to some folks. We welcome everyone, so rather than take chances this change should eliminate any confusion. Also "green" can be used to mean "not experienced or trained" which is _exactly_ what we're going for, so fits perfectly.